Remove node-wide clause filter, simplify clause sharing digestion code

+++ src/worker.cpp
@@ -258,7 +258,7 @@ void Worker::mainProgram() {
             float* result = _sys_state.getGlobal();
             int verb = (_world_rank == 0 ? Console::INFO : Console::VVVVERB);
             Console::log(verb, "sysstate busyratio=%.3f jobs=%i globmem=%.2fGB hops=%i", 
-                        result[0]/MyMpi::size(_comm), (int)result[1], result[2], result[3]);
+                        result[0]/MyMpi::size(_comm), (int)result[1], result[2], (int)result[3]);
             _sys_state.setLocal(SYSSTATE_NUMHOPS, 0); // reset #hops
         }
